+++ src/main/java/com/hta2405/unite/dto/VacationInfoDTO.java
@@ -10,6 +10,7 @@ import java.util.List;
 @AllArgsConstructor
 @Builder
 public class VacationInfoDTO {
+    private String targetEmpId; //직원 id
     private String targetEmpName; //직원이름
     private String jobName; //직책명
     private List<VacationDTO> vacationDTOList; //특정 년도 휴가정보
