+++ src/main/java/com/hta2405/unite/service/AttendService.java
@@ -252,6 +252,7 @@ public class AttendService {
                 .sum();
 
         return VacationInfoDTO.builder()
+                .targetEmpId(emp.getEmpId())
                 .targetEmpName(emp.getEname())
                 .jobName(jobName)
                 .vacationCount(emp.getVacationCount())
