+++ src/main/resources/templates/attend/attendBox.html
@@ -53,7 +53,8 @@
         <button id="startButton" class="btn button-active btn-sm">출근하기</button>
         <button id="endButton" class="btn button-inactive btn-sm">퇴근하기</button>
     </div>
-    <script th:inline="javascript">
+    <script>
+        const contextPath = /*[[@{/}]]*/ '';
         // 오늘 날짜 표시
         $(document).ready(function () {
             const now = new Date();
@@ -66,7 +67,7 @@
 
             // 페이지 로딩 시 AJAX로 출근 기록 확인
             $.ajax({
-                url: /*[[@{/attend/record}]]*/ '', // 서버에서 출근 기록을 가져오는 URL
+                url: contextPath + '/api/attend/record', // 서버에서 출근 기록을 가져오는 URL
                 method: "GET",
                 dataType: "json",
                 success: function (data) {
@@ -175,7 +176,7 @@
 
             // 출근 시간 저장을 위한 AJAX 요청
             $.ajax({
-                url: /*[[@{/attend/in}]]*/'',
+                url: contextPath + '/api/attend/in',
                 method: "POST",
                 data: {attendType: type},
                 success: function () {
@@ -198,7 +199,7 @@
             }
             // 퇴근 시간 저장을 위한 AJAX 요청
             $.ajax({
-                url: /*[[@{/attend/out}]]*/'',
+                url: contextPath + '/api/attend/out',
                 method: "POST",
                 success: function () {
                     alert("퇴근하였습니다.")
